refactor: Move some code from core.js
